backval

Calculate backgroud values for vegetation index.

backval(y, t, w, Tn, minT = 5, nptperyear, ...)

Arguments

y
Numeric vector, vegetation index time-series

t
Numeric vector, Date variable

w
(optional) Numeric vector, weights of y. If not specified, weights of all NA values will be wmin, the others will be 1.0.

Tn
Numeric vector, night temperature, default is null. If provided, Tn is used to help divide ungrowing period, and then get background value in ungrowing season (see details in backval()).

minT
min temperature for growing season.

nptperyear
Integer, number of images per year.

...
Others will be ignored.

Value

back If back value is NA, it is impossible to extract phenology here.

Details

Night temperature Tn >= Tmin (default 5 degree) defined as raw growing season. Background value is determined from two neighboring vegetation in raw growing season by assuming that the background and vegetation abundance could remain the same during a consecutive two yearperiod. Details can be seen in Zhang et al., (2015).

Note

This function only works in every growing season.

References

Zhang, X., 2015. Reconstruction of a complete global time series of daily vegetation index trajectory from long-term AVHRR data. Remote Sens. Environ. 156, 457–472. https://doi.org/10.1016/j.rse.2014.10.012.

Zhang, Y., Xiao, X., Jin, C., Dong, J., Zhou, S., Wagle, P., Joiner, J., Guanter, L., Zhang, Y., Zhang, G., Qin, Y., Wang, J., Moore, B., 2016. Consistency between sun-induced chlorophyll fluorescence and gross primary production of vegetation in North America. Remote Sens. Environ. 183, 154–169. https://doi.org/10.1016/j.rse.2016.05.015.