Double logistics functions

Define double logistics, piecewise logistics and many other functions to curve fit VI time-series

Logistic The traditional simplest logistic function. It can be only used in half growing season, i.e. vegetation green-up or senescence period.

doubleLog.Zhang Piecewise logistics, (Zhang Xiaoyang, RSE, 2003).

doubleAG Asymmetric Gaussian.

doubleLog.Beck Beck logistics.

doubleLog.Gu Gu logistics.

doubleLog.Elmore Elmore logistics.

doubleLog.Klos Klos logistics.

Logistic(par, t)

doubleLog.Zhang(par, t)

doubleLog.AG(par, t)

doubleLog.AG2(par, t)

doubleLog.Beck(par, t)

doubleLog.Elmore(par, t)

doubleLog.Gu(par, t)

doubleLog.Klos(par, t)

Arguments

par
A vector of parameters

t
A Date or numeric vector

Details

All of those function have par and formula attributes for the convenience for analytical D1 and D2

References

Peter M. Atkinson, et al., 2012, RSE, 123:400-417