Weight updating functions

wSELF weigth are not changed and return the original.

wTSM weight updating method in TIMESAT.

wBisquare Bisquare weight update method. wBisquare has been modified to emphasis on upper envelope.

wBisquare0 Traditional Bisquare weight update method.

wChen Chen et al., (2004) weight updating method.

wBeck Beck et al., (2006) weigth updating method. wBeck need sos and eos input. The function parameter is different from others. It is still not finished.

wSELF(y, yfit, w, ...)

wTSM(y, yfit, w, iter = 2, nptperyear, wfact = 0.5, ...)

wBisquare0(y, yfit, w, ..., wmin = 0.2)

wBisquare(y, yfit, w, ..., wmin = 0.2, .toUpper = TRUE)

wChen(y, yfit, w, ..., wmin = 0.2)

wKong(y, yfit, w, ..., wmin = 0.2)

Arguments

y

Numeric vector, vegetation index time-series

yfit

Numeric vector curve fitting values.

w

(optional) Numeric vector, weights of y. If not specified, weights of all NA values will be wmin, the others will be 1.0.

...

other parameters are ignored.

iter

iteration of curve fitting.

nptperyear

Integer, number of images per year.

wfact

weight adaptation factor (0-1), equal to the reciprocal of 'Adaptation strength' in TIMESAT.

wmin

Double, minimum weight of bad points, which could be smaller the weight of snow, ice and cloud.

.toUpper

Boolean. Whether to approach the upper envelope?

Value

wnew Numeric Vector, adjusted weights.

References

Per J\"onsson, P., Eklundh, L., 2004. TIMESAT - A program for analyzing time-series of satellite sensor data. Comput. Geosci. 30, 833-845. https://doi.org/10.1016/j.cageo.2004.05.006.

https://au.mathworks.com/help/curvefit/smoothing-data.html#bq_6ys3-3

Garcia, D., 2010. Robust smoothing of gridded data in one and higher dimensions with missing values. Computational statistics & data analysis, 54(4), pp.1167-1178.

Chen, J., J\"onsson, P., Tamura, M., Gu, Z., Matsushita, B., Eklundh, L., 2004. A simple method for reconstructing a high-quality NDVI time-series data set based on the Savitzky-Golay filter. Remote Sens. Environ. 91, 332-344. https://doi.org/10.1016/j.rse.2004.03.014.

Beck, P.S.A., Atzberger, C., Hogda, K.A., Johansen, B., Skidmore, A.K., 2006. Improved monitoring of vegetation dynamics at very high latitudes: A new method using MODIS NDVI. Remote Sens. Environ. https://doi.org/10.1016/j.rse.2005.10.021

https://github.com/kongdd/phenopix/blob/master/R/FitDoubleLogBeck.R

Author

wTSM is implemented by Per J\"onsson, Malm\"o University, Sweden per.jonsson@ts.mah.se and Lars Eklundh, Lund University, Sweden lars.eklundh@nateko.lu.se. And Translated into Rcpp by Dongdong Kong, 01 May 2018.